@@ -1,98 +1,159 @@
-import React, { Component } from 'react';
+import React, { useState } from "react";
+import emailjs from "emailjs-com";
 
-class Contact extends Component {
-  render() {
+const Contact = ({ data }) => {
+  const [name, setName] = useState("");
+  const [email, setEmail] = useState("");
+  const [subject, setSubject] = useState("");
+  const [message, setMessage] = useState("");
 
-    if(this.props.data){
-      var name = this.props.data.name
-      var city = this.props.data.address.city;
-      var state = this.props.data.address.state;
-      var zip = this.props.data.address.zip;
-      var phone= this.props.data.phone;
-      var email = this.props.data.email;
-      var message = this.props.data.contactmessage;
-    }
-
-    return (
-      <section id="contact">
-
-         <div className="row section-head">
-
-            <div className="two columns header-col">
-
-               <h1><span>Get In Touch.</span></h1>
-
-            </div>
-
-            <div className="ten columns">
-
-                  <p className="lead">{message}</p>
-                  <br/>
-                  <p className="lead warning">Warning: Contact form not yet configured, please contact me via normal email for now</p>
-
-            </div>
-
-         </div>
-
-         <div className="row">
-            <div className="eight columns">
-
-               <form action="" method="post" id="contactForm" name="contactForm">
-					<fieldset>
-
-                  <div>
-						   <label htmlFor="contactName">Name <span className="required">*</span></label>
-						   <input type="text" defaultValue="" size="35" id="contactName" name="contactName" onChange={this.handleChange}/>/>
-                  </div>
-
-                  <div>
-						   <label htmlFor="contactEmail">Email <span className="required">*</span></label>
-						   <input type="text" defaultValue="" size="35" id="contactEmail" name="contactEmail" onChange={this.handleChange}/>/>
-                  </div>
-
-                  <div>
-						   <label htmlFor="contactSubject">Subject</label>
-						   <input type="text" defaultValue="" size="35" id="contactSubject" name="contactSubject" onChange={this.handleChange}/>/>
-                  </div>
-
-                  <div>
-                     <label htmlFor="contactMessage">Message <span className="required">*</span></label>
-                     <textarea cols="50" rows="15" id="contactMessage" name="contactMessage"></textarea>
-                  </div>
-
-                  <div>
-                     <button className="submit">Submit</button>
-                     <span id="image-loader">
-                        <img alt="" src="images/loader.gif" />
-                     </span>
-                  </div>
-					</fieldset>
-				   </form>
-
-           <div id="message-warning"> Error boy</div>
-				   <div id="message-success">
-                  <i className="fa fa-check"></i>Your message was sent, thank you!<br />
-				   </div>
-           </div>
-
-
-            <aside className="four columns footer-widgets">
-               <div className="widget widget_contact">
-
-					   <h4>Address and Phone</h4>
-					   <p className="address">
-						   {name}<br />
-						   {city} <br />
-						   {state}, {zip}<br />
-						   <span>{phone}</span><br />
-               <span><a href="mailto:tbakerx@gmail.com">{email}</a></span>
-					   </p>
-				   </div>
-            </aside>
-      </div>
-   </section>
-    );
+  if (data) {
+    var contactName = data.name;
+    var street = data.address.street;
+    var city = data.address.city;
+    var state = data.address.state;
+    var zip = data.address.zip;
+    var phone = data.phone;
+    var contactEmail = data.email;
+    var contactMessage = data.contactmessage;
   }
-}
+
+  /* const submitForm = () => {
+    window.open(
+      `mailto:${contactEmail}?subject=${encodeURIComponent(
+        subject
+      )}&body=${encodeURIComponent(name)} (${encodeURIComponent(
+        email
+      )}): ${encodeURIComponent(message)}`
+    );
+  }; */
+
+  function sendEmail(e){
+    e.preventDefault();
+
+    emailjs.sendForm('service_hdwfdne', 'template_o1azqpf', e.target, 'user_mwMvWxsbCwH7eDldtVkhc')
+      .then((result) => {
+          console.log(result.text);
+      }, (error) => {
+          console.log(error.text);
+      });
+      window.alert("Email sent successfully");
+      window.location.reload();
+  }
+
+  return (
+    <section id="contact">
+      <div className="row section-head">
+        <div className="two columns header-col">
+          <h1>
+            <span>Get In Touch.</span>
+          </h1>
+        </div>
+
+        <div className="ten columns">
+          <p className="lead">{contactMessage}</p>
+        </div>
+      </div>
+
+      <div className="row">
+        <div className="eight columns">
+          <form onSubmit={sendEmail}>
+            <fieldset>
+              <div>
+                <label htmlFor="contactName">
+                  Name <span className="required">*</span>
+                </label>
+                <input
+                  type="text"
+                  defaultValue=""
+                  value={name}
+                  size="35"
+                  id="contactName"
+                  name="name"
+                  onChange={(e) => setName(e.target.value)}
+                />
+              </div>
+
+              <div>
+                <label htmlFor="contactEmail">
+                  Email <span className="required">*</span>
+                </label>
+                <input
+                  type="text"
+                  defaultValue=""
+                  value={email}
+                  size="35"
+                  id="contactEmail"
+                  name="email"
+                  onChange={(e) => setEmail(e.target.value)}
+                />
+              </div>
+
+              <div>
+                <label htmlFor="contactSubject">Subject</label>
+                <input
+                  type="text"
+                  defaultValue=""
+                  value={subject}
+                  size="35"
+                  id="contactSubject"
+                  name="subject"
+                  onChange={(e) => setSubject(e.target.value)}
+                />
+              </div>
+
+              <div>
+                <label htmlFor="contactMessage">
+                  Message <span className="required">*</span>
+                </label>
+                <textarea
+                  cols="50"
+                  rows="15"
+                  value={message}
+                  onChange={(e) => setMessage(e.target.value)}
+                  id="contactMessage"
+                  name="message"
+                ></textarea>
+              </div>
+
+              <div>
+                <button type="submit" className="submit">Submit</button>
+{/* 
+                <button type="submit" onClick={handleClick}>Send mail</button>
+                 */}
+              </div>
+            </fieldset>
+          </form>
+
+          <div id="message-warning"> Error boy</div>
+          <div id="message-success">
+            <i className="fa fa-check"></i>Your message was sent, thank you!
+            <br />
+          </div>
+        </div>
+
+        <aside className="four columns footer-widgets">
+          <div className="widget widget_contact">
+            <h4>Address and Phone</h4>
+            <p className="address">
+              {contactName}
+              <br />
+              {contactEmail}
+              <br />
+              <br />
+              {street} <br />
+              {city}, {state} {zip}
+              <br />
+              <span>{phone}</span>
+            </p>
+          </div>
+        </aside>
+      </div>
+    </section>
+  );
+};
 
 export default Contact;
+/* 
+I would wish to thank you God for enabling me to make my first contact form, all this has been your sufficient grace. Praise the Lord oh my soul! To many more websites.
+ */
