@@ -1,75 +1,77 @@
-import React, { Component } from "react";
+import React from "react";
+import TypeWriter from "react-typewriter";
 
-class Header extends Component {
-  render() {
-    if (this.props.data) {
-      var name = this.props.data.name;
-      var occupation = this.props.data.occupation;
-      var description = this.props.data.description;
-      var jobDescription = this.props.data.jobDescription;
-      var employer = this.props.data.employer;
-      var city = this.props.data.address.city;
-      var networks = this.props.data.social.map(function (network) {
-        return (
-          <li key={network.name}>
-            <a href={network.url}>
-              <i className={network.className}></i>
+const Header = ({ data }) => {
+  if (data) {
+    var name = data.name;
+    var occupation = data.occupation;
+    var description = data.description;
+    var city = data.address.city;
+    var networks = data.social.map(function (network) {
+      return (
+        <li key={network.name}>
+          <a href={network.url}>
+            <i className={network.className}></i>
+          </a>
+        </li>
+      );
+    });
+  }
+
+  return (
+    <header id="home">
+      <nav id="nav-wrap">
+        <a className="mobile-btn" href="#nav-wrap" title="Show navigation">
+          Show navigation
+        </a>
+        <a className="mobile-btn" href="#home" title="Hide navigation">
+          Hide navigation
+        </a>
+
+        <ul id="nav" className="nav">
+          <li className="current">
+            <a className="smoothscroll" href="#home">
+              Home
             </a>
           </li>
-        );
-      });
-    }
+          <li>
+            <a className="smoothscroll" href="#about">
+              About
+            </a>
+          </li>
+          <li>
+            <a className="smoothscroll" href="#resume">
+              Resume
+            </a>
+          </li>
+          <li>
+            <a className="smoothscroll" href="#contact">
+              Contact
+            </a>
+          </li>
+        </ul>
+      </nav>
 
-    return (
-      <header id="home">
-        <nav id="nav-wrap">
-          <a className="mobile-btn" href="#nav-wrap" title="Show navigation">
-            Show navigation
-          </a>
-          <a className="mobile-btn" href="#home" title="Hide navigation">
-            Hide navigation
-          </a>
-
-          <ul id="nav" className="nav">
-            <li className="current">
-              <a className="smoothscroll" href="#home">
-                Home
-              </a>
-            </li>
-            <li>
-              <a className="smoothscroll" href="#about">
-                About
-              </a>
-            </li>
-            <li>
-              <a className="smoothscroll" href="#contact">
-                Contact
-              </a>
-            </li>
-          </ul>
-        </nav>
-
-        <div className="row banner">
-          <div className="banner-text">
-            <h1 className="responsive-headline">I'm {name}</h1>
-            <h3>
-              I'm a <span>{city}</span> based <span>{occupation}</span>.
-              {description}
-              <span>{employer}</span><span> {jobDescription}</span>.
-            </h3>
-            <hr />
-            <ul className="social">{networks}</ul>
-          </div>
+      <div className="row banner">
+        <div className="banner-text">
+          <h1 className="responsive-headline">
+            <TypeWriter typing={0.5}>{name ? `I'm ${name}.` : null}</TypeWriter>
+          </h1>
+          <h3>
+            I'm a <span>{city}</span> based <span>{occupation}</span>. {description}.
+          </h3>
+          <hr />
+          <ul className="social">{networks}</ul>
         </div>
+      </div>
 
-        <p className="scrolldown">
-          <a className="smoothscroll" href="#about">
-            <i className="icon-down-circle"></i>
-          </a>
-        </p>
-      </header>
-    );
-  }
-}
+      <p className="scrolldown">
+        <a className="smoothscroll" href="#about">
+          <i className="icon-down-circle"></i>
+        </a>
+      </p>
+    </header>
+  );
+};
 
 export default Header;
